web: re-use vars during template interpolation if same path is requested multiple times
This commit updates the `vars` module to only call `Get` once for each variable referenced during interpolation. If a variable is referenced multiple times, the data will be cached and each sub-field would be retrieved from this cached data instead of making separate calls to the credential manager (or local vars, or var source).

As an example, say there is the following task configuration in a pipeline...

```
platform: linux
[...]

params:
  AWS_ACCESS_KEY_ID: ((aws-credentials.access_key))
  AWS_SECRET_ACCESS_KEY: ((aws-credentials.secret_key))
```

Prior to this change, the `aws-credentials` variable would be looked up twice, once while resolving the `access_key` sub-field and again for the `secret_key` sub-field. This would be subject to a race condition should the value of `aws-credentials` change during interpolation, resulting in a mismatching access key / secret key pair.

The specific use case we are looking to fix is when using a Vault var source to provide dynamically provisioned AWS credentials to a pipeline. In this case, Vault provides a unique access key / secret key pair each time the variable is retrieved, and therefore there is a need to re-use the same variable during interpolation.