web: add `eventual_consistency_delay` field to Vault configuration
When reading IAM credentials from Vault, they can be returned before they are ready to be used. This seems to be a quirk of AWS and eventual consistency.

There is some discussion in an issue on Vault's Github repository on possible workarounds, but nothing has come from it.
https://github.com/hashicorp/vault/issues/3115

This commit adds a field, called `eventual_consistency_delay`, which introduces an artificial delay when reading a secret from Vault in an attempt to allow time for the credentials to become valid. While not an ideal solution, I couldn't think of a better approach, especially when IAM credentials could be used early in a job before there is a chance to introduce a delay via other means (eg. a task step that just sleeps).